Capita Selecta in Cyber Security/Assignments/Assignment_1/requirements.txt
geoip2==4.5.0
python_dateutil==2.8.2
tld==0.12.6
